backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.1
ultralytics==8.0.220
torch>=2.0.0
torchvision>=0.15.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4